oci-unpack: created project to fetch oci bundles
For our containers we need OCI Runtime Bundles. Those contain a rootfs
and a config.json. From those we can easily create the config.json
(using `runc spec` or `crun spec`). But creating a rootfs is not
trivial: we need a base OCI Image and unpack it.

So far I did not find a rust library that implements unpacking, so this
project was created for that purpose. Downloading images from an OCI
Registry will be done using the `oci-client` crate.