warpfoge-executor: added runtime argument and removed redundant gvisor
Using gvisor as a container runtime should be simply possible by
calling: `warpforge run --runtime /path/to/gvisor <target>`

If we want to add exceptions for how we configure our containers for
different runtimes, we can add that later. For now it's simpler to just
work with one single executor module.